backend/requirements.txt
# Byte in Bottle Backend - Development Dependencies
# Generated from uv project

fastapi>=0.120.0
ollama>=0.6.0
python-dotenv>=1.1.1
uvicorn>=0.38.0
